organize burst detection into own folder
