+++ bycycle/burst/__init__.py
@@ -1,0 +1,5 @@
+"""Burst detection."""
+
+from .amp import detect_bursts_amp
+from .cycle import detect_bursts_cycles
+from .dualthresh import detect_bursts_dual_threshold
