+++ bycycle/burst/dualthresh.py
@@ -1,0 +1,4 @@
+"""Detect bursts: dual threshold approach."""
+
+# This approach is currently aliased in from NeuroDSP
+from neurodsp.burst import detect_bursts_dual_threshold
